Deal with params that are not a map correctly
The params argument for `Phoenix.Channel.join/3` is
not guaranteed to be a map. It is whatever is passed
as the payload value in the request.

The following payload lead to errors:

```
{"topic":"users:11187562","event":"phx_join","payload":"","ref":"1"}
```

This change fixes that and makes sure the instrumentation
code works also for params that are not a Map.